ドッグランへの入退場記録処理の作成 (#156)
* ドッグランへの入場記録処理の作成(#101)

* ドッグランへの退場記録処理の作成(#101)

* 今日のチェックイン済みドッグランの取得処理作成.clain内roleをintに変更(#101)

---------

+++ internal/auth/middleware/jwt.go
@@ -167,7 +167,7 @@ func (aj *authJwt) jwtIDValid(c echo.Context, ac *handler.AccountClaims) error {
 	}
 
 	// Roleによる設定分岐
-	getJwtID := func(role int64, id int64) (string, error) {
+	getJwtID := func(role int, id int64) (string, error) {
 		switch role {
 		// dogowner
 		case handler.DOGOWNER_ROLE:
